backend/requirements.txt
Flask==2.2.5
Flask_Cors==5.0.0
protobuf==6.33.2
pymongo==3.10.1
python-dotenv==1.2.1
